Create StellarPayment and StellarTransaction add them as part of the Sep31Transaction.
