+++ api-schema/src/main/java/org/stellar/anchor/api/shared/StellarPayment.java
@@ -1,0 +1,12 @@
+package org.stellar.anchor.api.shared;
+
+import lombok.Data;
+
+@Data
+public class StellarPayment {
+  String id;
+  String paymentType;
+  String sourceAccount;
+  String destinationAccount;
+  Amount amount;
+}
