+++ platform/build.gradle.kts
@@ -28,6 +28,7 @@ dependencies {
   implementation(libs.commons.cli)
   implementation(libs.commons.io)
   implementation(libs.flyway.core)
+  implementation(libs.hibernate.types)
   implementation(libs.google.gson)
   implementation(libs.java.stellar.sdk)
